# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Blog",
        "description": "Cette API permet de gérer un blog.",
        "contact": {
            "name": "Joséphin Sylvère",
            "email": "josephinsylvere@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Récupère la liste de tous les posts",
                "description": "Retourne une liste de tous les posts avec leurs détails",
                "operationId": "getAllPosts",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Crée un nouveau post",
                "description": "Enregistre un nouveau post avec les données fournies après validation.",
                "operationId": "createPost",
                "requestBody": {
                    "description": "Données du nouveau post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Récupère un post par son ID",
                "description": "Retourne un post et ses relations (catégories, tags, utilisateur) par son ID.",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Met à jour un post existant",
                "description": "Met à jour un post existant avec les données fournies après validation.",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données mises à jour du post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Supprime un post",
                "description": "Supprime un post spécifié par son ID et renvoie un message de succès.",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le post a été supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/posts/{id}/categories": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Récupère les catégories d'un post",
                "description": "Retourne les catégories associées à un post spécifique.",
                "operationId": "getPostCategories",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/posts/{id}/tags": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Récupère les tags d'un post",
                "description": "Retourne les tags associés à un post spécifique.",
                "operationId": "getPostTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Categorie",
                "description": "Un modèle de catégorie pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique de la catégorie",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Le nom de la catégorie",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Le slug de la catégorie",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création de la catégorie",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour de la catégorie",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "title": "Tag",
                "description": "Un modèle de tag pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique du tag",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Le nom du tag",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Le slug du tag",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création du tag",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour du tag",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "title": "Post",
                "description": "Un modèle de post pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique du post",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Le titre du post",
                        "type": "string"
                    },
                    "content": {
                        "description": "Le contenu du post",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "L'identifiant de l'utilisateur qui a créé le post",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création du post",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour du post",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "auth_bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le jeton JWT de cette manière: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}